organized api directory structure
