+++ src/app/api/auth/register/aina/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { db } from '../../../../../db/kysely/client';
+import { db } from '../../../../../../db/kysely/client';
 import { getUserID } from '@/lib/server-utils';
 
 export async function POST(request: Request) {
