+++ src/app/api/auth/register/route.ts
